refactor(aiProvider): generateText uses setStickyFallback helper
