+++ backend/src/aiProvider/registry.js
@@ -1,41 +1,297 @@
 /**
  * @module aiProvider/registry
- * @description Public re-export facade for provider-state operations.
+ * @description Owns mutable provider state — canonical state owner per AI-002.
  *
- * The actual state lives in `./index.js` (orchestrator) so this file stays
- * tiny and stable. The facade exists per AI-002 spec to give every consumer
- * a single canonical import path, which is what lets future AI-005
- * multi-agent dispatch evolve the breaker keyspace, key-resolution, or
- * detection logic without breaking any caller — they all already import
- * from `./registry.js`.
+ * The orchestrator (`./index.js`) imports from here; the reverse is forbidden.
+ * Future AI-005 multi-agent dispatch can extend the breaker keyspace
+ * (`breakerKey(provider, role)`) inside this file without touching consumers.
  *
- * Why not physically relocate state here in this PR: doing so requires
- * touching every internal call site in `index.js` (~40+ call sites) which
- * would balloon the diff past the reviewable size REVIEW.md asks for.
- * The facade locks the import surface; relocation is a pure mechanical
- * move-and-shadow operation that can land in a follow-up `chore` PR
- * without affecting any consumer — `git mv`-shape, not a redesign.
- *
- * Per AGENT.md "every finding produces an outcome": the relocation is
- * tracked in ROADMAP.md as AI-002b.
+ * State owned here: runtime key store, Ollama runtime config, active-provider
+ * override, sticky fallback, per-provider circuit breakers, compat-slot
+ * config (TTL-cached), provider detection + usability + key resolution.
  */
-export {
-  // Mutators
-  setRuntimeKey,
-  setRuntimeOllama,
-  setActiveProvider,
-  loadKeysFromDatabase,
+import * as apiKeyRepo from "../database/repositories/apiKeyRepo.js";
+import * as compatConfigCache from "../utils/compatConfigCache.js";
+import { formatLogLine } from "../utils/logFormatter.js";
+import { CLOUD_KEY_MAP, CLOUD_DETECT_ORDER } from "./modelCatalog.js";
 
-  // Detection + status
-  getProvider,
-  hasProvider,
-  isLocalProvider,
-  isProviderDegraded,
-  getProviderMeta,
-  getProviderName,
-  getConfiguredKeys,
-  getSupportedProviders,
+// ── Mutable state ────────────────────────────────────────────────────────────
+const runtimeKeys = {};
+let runtimeOllamaBaseUrl = "";
+let runtimeOllamaModel   = "";
+let runtimeOllamaDisabled = false;
+let runtimeActiveProvider = null;
+let _stickyFallbackProvider = null;
+let _stickyFallbackExpiry   = 0;
 
-  // Connectivity check (Ollama)
-  checkOllamaConnection,
-} from "./index.js";
+export const STICKY_FALLBACK_TTL_MS = 10 * 60 * 1000;
+const CIRCUIT_BREAKER_THRESHOLD = 1;
+const CIRCUIT_BREAKER_COOLDOWN_MS = 5 * 60 * 1000;
+
+/** @type {Object<string, {failures: number, disabledUntil: number}>} */
+const circuitBreakers = {};
+
+// ── Compat helpers ───────────────────────────────────────────────────────────
+export function isCompatProvider(provider) {
+  return typeof provider === "string"
+    && provider.startsWith("compat:")
+    && provider.length > "compat:".length;
+}
+
+export function getCompatConfig(provider) {
+  if (!isCompatProvider(provider)) return null;
+  // TTL cache avoids hitting SQLite (decrypt + JSON.parse) on every AI call.
+  return compatConfigCache.get(provider, () => apiKeyRepo.getCompatSlot(provider));
+}
+
+// ── Key resolution ───────────────────────────────────────────────────────────
+export function getKey(envName) {
+  // `in` semantics so runtimeKeys[envName] = "" (explicit deactivation) takes
+  // precedence over the env var. `||` would treat "" as falsy.
+  if (envName in runtimeKeys) return runtimeKeys[envName];
+  const envVal = process.env[envName] || "";
+  if (envVal) return envVal;
+  // DEMO-MODE: platform-owned demo key for Google when no user key is set.
+  if (envName === "GOOGLE_API_KEY" && process.env.DEMO_GOOGLE_API_KEY) {
+    return process.env.DEMO_GOOGLE_API_KEY;
+  }
+  return "";
+}
+
+/** Get a user-configured key WITHOUT the demo fallback (for BYOK detection). */
+export function getUserConfiguredKey(envName) {
+  if (envName in runtimeKeys) return runtimeKeys[envName];
+  return process.env[envName] || "";
+}
+
+export function getOllamaBaseUrl() {
+  return runtimeOllamaBaseUrl || process.env.OLLAMA_BASE_URL || "http://localhost:11434";
+}
+
+export function getOllamaModel() {
+  return runtimeOllamaModel || process.env.OLLAMA_MODEL || "mistral:7b";
+}
+
+export function hasOllamaConfig() {
+  return !!(runtimeOllamaBaseUrl || runtimeOllamaModel
+    || process.env.OLLAMA_BASE_URL || process.env.OLLAMA_MODEL);
+}
+
+export function isOllamaDisabled() { return runtimeOllamaDisabled; }
+
+// loadKeysFromDatabase seeds caches from DB at startup without clobbering env.
+export function setCloudKeyFromDb(envName, value) {
+  if (!process.env[envName] && !(envName in runtimeKeys)) {
+    runtimeKeys[envName] = String(value);
+    return true;
+  }
+  return false;
+}
+
+export function setOllamaCacheFromDb(cfg) {
+  if (!runtimeOllamaBaseUrl && !process.env.OLLAMA_BASE_URL) {
+    runtimeOllamaBaseUrl = cfg.baseUrl || "";
+  }
+  if (!runtimeOllamaModel && !process.env.OLLAMA_MODEL) {
+    runtimeOllamaModel = cfg.model || "";
+  }
+  runtimeOllamaDisabled = false;
+}
+
+// ── Mutators ─────────────────────────────────────────────────────────────────
+export function setRuntimeKey(provider, key) {
+  if (isCompatProvider(provider)) {
+    // Compat providers persist via apiKeyRepo.setCompatSlot() in settings.js.
+    resetCircuitBreaker(provider);
+    clearStickyFallback();
+    return;
+  }
+  const envName = CLOUD_KEY_MAP[provider];
+  if (!envName) return;
+  runtimeKeys[envName] = key;
+  // FEA-003: reset breaker so new credentials are retried immediately.
+  resetCircuitBreaker(provider);
+  clearStickyFallback();
+  try {
+    if (key) apiKeyRepo.set(provider, key);
+    else apiKeyRepo.remove(provider);
+  } catch (err) {
+    // DB unavailable during tests or before init — safe to ignore.
+    console.error(formatLogLine("error", null, `[aiProvider] Failed to persist key for ${provider}: ${err.message}`));
+  }
+}
+
+export function setRuntimeOllama({ baseUrl, model, disabled } = {}) {
+  if (baseUrl  !== undefined) runtimeOllamaBaseUrl  = baseUrl;
+  if (model    !== undefined) runtimeOllamaModel    = model;
+  if (disabled !== undefined) runtimeOllamaDisabled = disabled;
+  try {
+    if (disabled) {
+      apiKeyRepo.remove("local");
+    } else if (runtimeOllamaBaseUrl || runtimeOllamaModel) {
+      apiKeyRepo.set("local", { baseUrl: runtimeOllamaBaseUrl, model: runtimeOllamaModel });
+    }
+  } catch (err) {
+    console.error(formatLogLine("error", null, `[aiProvider] Failed to persist Ollama config: ${err.message}`));
+  }
+}
+
+export function setActiveProvider(provider) {
+  runtimeActiveProvider = provider || null;
+  // User chose a provider — clear sticky fallback so detection re-evaluates.
+  clearStickyFallback();
+}
+
+// ── Sticky fallback ──────────────────────────────────────────────────────────
+export function setStickyFallback(provider) {
+  _stickyFallbackProvider = provider;
+  _stickyFallbackExpiry   = Date.now() + STICKY_FALLBACK_TTL_MS;
+}
+
+export function clearStickyFallback() {
+  _stickyFallbackProvider = null;
+  _stickyFallbackExpiry   = 0;
+}
+
+export function stickyFallbackActive() {
+  return !!(_stickyFallbackProvider && Date.now() < _stickyFallbackExpiry);
+}
+
+// ── Circuit breaker (FEA-003) ────────────────────────────────────────────────
+export function recordProviderFailure(provider) {
+  if (!circuitBreakers[provider]) circuitBreakers[provider] = { failures: 0, disabledUntil: 0 };
+  circuitBreakers[provider].failures += 1;
+  if (circuitBreakers[provider].failures >= CIRCUIT_BREAKER_THRESHOLD) {
+    circuitBreakers[provider].disabledUntil = Date.now() + CIRCUIT_BREAKER_COOLDOWN_MS;
+    console.warn(formatLogLine("warn", null, `[aiProvider] Circuit breaker tripped for ${provider} — disabled for ${CIRCUIT_BREAKER_COOLDOWN_MS / 1000}s after ${CIRCUIT_BREAKER_THRESHOLD} consecutive rate-limit failures`));
+  }
+}
+
+export function recordProviderSuccess(provider) {
+  if (circuitBreakers[provider]) circuitBreakers[provider].failures = 0;
+}
+
+export function isCircuitBreakerOpen(provider) {
+  const cb = circuitBreakers[provider];
+  if (!cb) return false;
+  if (cb.disabledUntil > Date.now()) return true;
+  if (cb.disabledUntil > 0) { cb.disabledUntil = 0; cb.failures = 0; }
+  return false;
+}
+
+function resetCircuitBreaker(provider) {
+  if (circuitBreakers[provider]) {
+    circuitBreakers[provider].failures = 0;
+    circuitBreakers[provider].disabledUntil = 0;
+  }
+}
+
+// ── Provider detection ───────────────────────────────────────────────────────
+export function isProviderUsable(provider) {
+  if (provider === "local") return !runtimeOllamaDisabled;
+  if (isCompatProvider(provider)) {
+    // Wrap so a transient DB failure during cache miss doesn't crash hot paths.
+    try {
+      const compat = getCompatConfig(provider);
+      return !!(compat?.apiKey && compat?.baseUrl && compat?.model);
+    } catch { return false; }
+  }
+  const envName = CLOUD_KEY_MAP[provider];
+  if (!envName) return false;
+  if (envName in runtimeKeys) return runtimeKeys[envName].length > 0;
+  if (process.env[envName]) return true;
+  if (envName === "GOOGLE_API_KEY" && process.env.DEMO_GOOGLE_API_KEY) return true;
+  return false;
+}
+
+export function detectProvider() {
+  // Sticky fallback first — a successful rate-limit fallback pins the working
+  // provider until the TTL expires, even if the user has the original
+  // (rate-limited) provider selected in the dropdown.
+  if (_stickyFallbackProvider && Date.now() < _stickyFallbackExpiry) {
+    if (isProviderUsable(_stickyFallbackProvider)) return _stickyFallbackProvider;
+    clearStickyFallback();
+  } else if (_stickyFallbackProvider) {
+    clearStickyFallback();
+  }
+
+  if (runtimeActiveProvider) {
+    if (isProviderUsable(runtimeActiveProvider)) return runtimeActiveProvider;
+    runtimeActiveProvider = null;
+  }
+
+  const forced = process.env.AI_PROVIDER?.toLowerCase();
+  if (forced) {
+    if (forced === "local") return "local";
+    if (!CLOUD_KEY_MAP[forced]) throw new Error(`Unknown AI_PROVIDER="${forced}". Valid: anthropic, openai, google, openrouter, local`);
+    if (!getKey(CLOUD_KEY_MAP[forced])) throw new Error(`AI_PROVIDER="${forced}" but ${CLOUD_KEY_MAP[forced]} is not set`);
+    return forced;
+  }
+
+  const detected = CLOUD_DETECT_ORDER.find((id) => isProviderUsable(id));
+  if (detected) return detected;
+
+  try {
+    const compatSlot = apiKeyRepo.listCompatSlots().find((id) => isProviderUsable(id));
+    if (compatSlot) return compatSlot;
+  } catch { /* DB unavailable — fall through to Ollama */ }
+
+  if (isProviderUsable("local") && hasOllamaConfig()) return "local";
+  return null;
+}
+
+export function getFallbackProviders(primaryProvider) {
+  if (primaryProvider === "local") return [];
+  // Cloud tier falls back to other cloud providers + compat slots — same wire
+  // format, same circuit-breaker accounting per slot. Local is excluded
+  // (cross-tier prompt-shape mismatch — see CLOUD_DETECT_ORDER doc).
+  let compatSlots = [];
+  try { compatSlots = apiKeyRepo.listCompatSlots(); } catch { /* DB unavailable */ }
+  const candidates = [...CLOUD_DETECT_ORDER, ...compatSlots];
+  return candidates.filter((p) =>
+    p !== primaryProvider
+    && isProviderUsable(p)
+    && !isCircuitBreakerOpen(p),
+  );
+}
+
+// ── Database key persistence ─────────────────────────────────────────────────
+/**
+ * Restore all persisted API keys and Ollama config from the DB into the
+ * runtime cache. Called once at server startup after the DB is initialised.
+ * Keys stored in the DB take precedence over default detection only when no
+ * matching env var is already set — env vars remain the canonical override.
+ *
+ * @returns {number} Number of providers successfully loaded from the database.
+ */
+export function loadKeysFromDatabase() {
+  let loaded = 0;
+  try {
+    const entries = apiKeyRepo.getAll();
+    for (const { provider, value } of entries) {
+      if (provider === "local") {
+        if (value && typeof value === "object") {
+          setOllamaCacheFromDb(value);
+          loaded += 1;
+        }
+      } else if (isCompatProvider(provider)) {
+        // Compat slots are read on demand via getCompatConfig() — no cache
+        // restore needed, just count for the boot log.
+        if (value && typeof value === "object" && value.apiKey && value.baseUrl && value.model) {
+          loaded += 1;
+        }
+      } else {
+        const envName = CLOUD_KEY_MAP[provider];
+        if (!envName) continue;
+        if (setCloudKeyFromDb(envName, value)) loaded += 1;
+      }
+    }
+    if (loaded > 0) {
+      console.log(formatLogLine("info", null, `[aiProvider] Restored ${loaded} provider key(s) from database`));
+    }
+  } catch (err) {
+    // Non-fatal: server still works with env vars; log and continue.
+    console.error(formatLogLine("error", null, `[aiProvider] Failed to load keys from database: ${err.message}`));
+  }
+  return loaded;
+}
